+++ CHANGELOG.rst
@@ -3,13 +3,14 @@ Changelog
 
 This document describes changes between each past release.
 
-0.8.0 (unreleased)
+0.7.3 (2016-07-27)
 ------------------
 
 **Bug fixes**
 
 - Fix signature inconsistency (timestamp) when several changes are sent from
-  the *source* to the *destination* collection (fixes #110)
+  the *source* to the *destination* collection.
+  Fixed ``e2e.py`` and ``validate_signature.py`` scripts (fixes #110)
 
 **Minor change**